// src/app/shared/models/user-profile.ts
export interface UserProfile {
  apiRequestToken: string;
  AutoTradeQuantity: number;
  apiKey: string;
  initialStopLoss: number;
  AutoTradeNumber: number;
  apiAccessToken: string;
  trailingStopLoss: number;
  userTestMode: boolean;
  IsAutoTradeEnabled: boolean;
  userName: string;
  IsEnabled: boolean;
  apiPublicToken: string;
  apiSecret: string;
  initialProfit: number;
  PauseMoneyManagement: boolean;
}